requirements.txt
pytest
pytest-bdd
selenium
python-dotenv
